requirements.txt
distro==1.5.0
numpy==1.20.2
pandas==1.2.4
pdftotext==2.1.5
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
tabula-py==2.2.0
